// app/services/types/home.ts
export interface UserProfileUpdateRequest {
  name: string;
  email: string;
  mobile: string;
}